Rename hub_session.identification to channel.
hub_session.identification was initiallly designed to identify the IP address of a hub.
It was needed when we were designing our own push server but since we moved to PubNub, we are using channel as identification.
So renaming the field.

Also renamed hub_session.Connect_time to connect_time for naming convention.