Added value_source to device_data and modified APIs accordingly.
Currently when a device(uPlug/uSense etc) sends data to cloud server through REST api.
This API just takes value which is not sufficient for multi function devices.
For example uSwitch has a current sensor, 4 triacs. But with current scheme we can send only one data and differentating them is difficult.

This change adds a source field(which is predefined), so that devices can include that to specify which data it is sending.